refactor: split scalex.scala into 6 source files and 4 test suites
Single-file implementation was 2660 lines, making it hard to navigate.
Split along existing section markers into focused modules under src/.

Source files: model.scala, extraction.scala, index.scala, analysis.scala, cli.scala, project.scala
Test suites: ExtractionSuite (47), IndexSuite (53), AnalysisSuite (20), CliSuite (56)

All 176 tests pass. No performance regression (benchmarked against scala3 repo).
