+++ src/cli.scala
@@ -1,1405 +1,7 @@
-//> using scala 3.8.2
-//> using dep org.scalameta::scalameta:4.15.2
-//> using dep com.google.guava:guava:33.5.0-jre
-
-import scala.meta.*
 import scala.collection.mutable
 import java.nio.file.{Files, Path}
-import java.io.{BufferedReader, BufferedInputStream, BufferedOutputStream,
-  DataInputStream, DataOutputStream, InputStreamReader}
-import java.util.concurrent.ConcurrentLinkedQueue
+import java.io.{BufferedReader, InputStreamReader}
 import scala.jdk.CollectionConverters.*
-import com.google.common.hash.{BloomFilter, Funnels}
-
-val ScalexVersion = "1.13.0"
-
-// ── Data types ──────────────────────────────────────────────────────────────
-
-enum SymbolKind(val id: Byte):
-  case Class     extends SymbolKind(0)
-  case Trait     extends SymbolKind(1)
-  case Object    extends SymbolKind(2)
-  case Def       extends SymbolKind(3)
-  case Val       extends SymbolKind(4)
-  case Var       extends SymbolKind(5)
-  case Type      extends SymbolKind(6)
-  case Enum      extends SymbolKind(7)
-  case Given     extends SymbolKind(8)
-  case Extension extends SymbolKind(9)
-  case Package   extends SymbolKind(10)
-
-object SymbolKind:
-  private val byId: Array[SymbolKind] = values.sortBy(_.id)
-  def fromId(id: Byte): SymbolKind = byId(id)
-
-case class SymbolInfo(
-    name: String,
-    kind: SymbolKind,
-    file: Path,
-    line: Int,
-    packageName: String,
-    parents: List[String] = Nil,
-    signature: String = "",
-    annotations: List[String] = Nil
-)
-
-case class Reference(file: Path, line: Int, contextLine: String, aliasInfo: Option[String] = None)
-case class GitFile(path: Path, oid: String)
-
-case class IndexedFile(
-    relativePath: String,
-    oid: String,
-    symbols: List[SymbolInfo],
-    identifierBloom: BloomFilter[CharSequence],
-    imports: List[String] = Nil,
-    aliases: Map[String, String] = Map.empty,
-    parseFailed: Boolean = false
-)
-
-enum RefCategory:
-  case Definition, ExtendedBy, ImportedBy, UsedAsType, Comment, Usage
-
-case class CategorizedRef(ref: Reference, category: RefCategory)
-
-enum Confidence:
-  case High, Medium, Low
-
-// ── Git ─────────────────────────────────────────────────────────────────────
-
-def gitLsFiles(workspace: Path): List[GitFile] =
-  val pb = ProcessBuilder("git", "ls-files", "--stage")
-  pb.directory(workspace.toFile)
-  pb.redirectErrorStream(true)
-  val proc = pb.start()
-  val reader = BufferedReader(InputStreamReader(proc.getInputStream))
-  val files = reader.lines().iterator().asScala.flatMap { line =>
-    val tabIdx = line.indexOf('\t')
-    if tabIdx < 0 then None
-    else
-      val parts = line.substring(0, tabIdx).split("\\s+")
-      val path = line.substring(tabIdx + 1)
-      if parts.length >= 2 && path.endsWith(".scala") then
-        Some(GitFile(workspace.resolve(path), parts(1)))
-      else None
-  }.toList
-  proc.waitFor()
-  files
-
-// ── Symbol extraction + bloom filter ────────────────────────────────────────
-
-def buildBloomFilterFromSource(source: String): BloomFilter[CharSequence] =
-  val expected = math.max(500, source.length / 15)
-  val bloom = BloomFilter.create(Funnels.unencodedCharsFunnel(), expected, 0.01)
-  var i = 0
-  val len = source.length
-  while i < len do
-    if source(i).isLetter || source(i) == '_' then
-      val start = i
-      while i < len && (source(i).isLetterOrDigit || source(i) == '_') do i += 1
-      val word = source.substring(start, i)
-      if word.length >= 2 then bloom.put(word)
-    else
-      i += 1
-  bloom
-
-private def extractParents(templ: Template): List[String] =
-  templ.inits.flatMap { init =>
-    init.tpe match
-      case Type.Name(name) => Some(name)
-      case Type.Select(_, Type.Name(name)) => Some(name)
-      case Type.Apply.After_4_6_0(Type.Name(name), _) => Some(name)
-      case Type.Apply.After_4_6_0(Type.Select(_, Type.Name(name)), _) => Some(name)
-      case _ => None
-  }
-
-private def buildSignature(name: String, kind: String, parents: List[String], tparams: List[String] = Nil): String =
-  val tps = if tparams.nonEmpty then tparams.mkString("[", ", ", "]") else ""
-  val ext = if parents.nonEmpty then s" extends ${parents.mkString(" with ")}" else ""
-  s"$kind $name$tps$ext"
-
-private def extractAnnotations(mods: List[Mod]): List[String] =
-  mods.collect { case Mod.Annot(init) =>
-    init.tpe match
-      case Type.Name(name) => name
-      case Type.Select(_, Type.Name(name)) => name
-      case _ => init.tpe.toString()
-  }
-
-private def extractImports(tree: Tree): (imports: List[String], aliases: Map[String, String]) =
-  val buf = mutable.ListBuffer.empty[String]
-  val aliases = mutable.Map.empty[String, String]
-  def visit(t: Tree): Unit =
-    t match
-      case i: Import =>
-        buf += i.toString()
-        i.importers.foreach { importer =>
-          importer.importees.foreach {
-            case r: Importee.Rename => aliases(r.name.value) = r.rename.value
-            case _ =>
-          }
-        }
-      case _ =>
-    t.children.foreach(visit)
-  visit(tree)
-  (buf.toList, aliases.toMap)
-
-def extractSymbols(file: Path): (symbols: List[SymbolInfo], bloom: BloomFilter[CharSequence], imports: List[String], aliases: Map[String, String], parseFailed: Boolean) =
-  val source = try Files.readString(file) catch
-    case _: Exception =>
-      val bloom = BloomFilter.create(Funnels.unencodedCharsFunnel(), 500, 0.01)
-      return (Nil, bloom, Nil, Map.empty, true)
-
-  val bloom = buildBloomFilterFromSource(source)
-
-  val input = Input.VirtualFile(file.toString, source)
-  val tree = try
-    given scala.meta.Dialect = scala.meta.dialects.Scala3
-    input.parse[Source].get
-  catch
-    case _: Exception =>
-      try
-        given scala.meta.Dialect = scala.meta.dialects.Scala213
-        input.parse[Source].get
-      catch
-        case _: Exception => return (Nil, bloom, Nil, Map.empty, true)
-
-  val pkg = tree.children.collectFirst { case p: Pkg => p.ref.toString() }.getOrElse("")
-  val (imports, aliases) = extractImports(tree)
-  val buf = mutable.ListBuffer.empty[SymbolInfo]
-
-  def visit(t: Tree): Unit = t match
-    case d: Defn.Class =>
-      val parents = extractParents(d.templ)
-      val tparams = d.tparamClause.values.map(_.name.value)
-      val sig = buildSignature(d.name.value, "class", parents, tparams)
-      val annots = extractAnnotations(d.mods)
-      buf += SymbolInfo(d.name.value, SymbolKind.Class, file, d.pos.startLine + 1, pkg, parents, sig, annots)
-    case d: Defn.Trait =>
-      val parents = extractParents(d.templ)
-      val tparams = d.tparamClause.values.map(_.name.value)
-      val sig = buildSignature(d.name.value, "trait", parents, tparams)
-      val annots = extractAnnotations(d.mods)
-      buf += SymbolInfo(d.name.value, SymbolKind.Trait, file, d.pos.startLine + 1, pkg, parents, sig, annots)
-    case d: Defn.Object =>
-      val parents = extractParents(d.templ)
-      val sig = buildSignature(d.name.value, "object", parents)
-      val annots = extractAnnotations(d.mods)
-      buf += SymbolInfo(d.name.value, SymbolKind.Object, file, d.pos.startLine + 1, pkg, parents, sig, annots)
-    case d: Pkg.Object =>
-      val parents = extractParents(d.templ)
-      val sig = buildSignature(d.name.value, "object", parents)
-      buf += SymbolInfo(d.name.value, SymbolKind.Object, file, d.pos.startLine + 1, pkg, parents, sig)
-    case d: Defn.Enum =>
-      val parents = extractParents(d.templ)
-      val tparams = d.tparamClause.values.map(_.name.value)
-      val sig = buildSignature(d.name.value, "enum", parents, tparams)
-      val annots = extractAnnotations(d.mods)
-      buf += SymbolInfo(d.name.value, SymbolKind.Enum, file, d.pos.startLine + 1, pkg, parents, sig, annots)
-    case d: Defn.Given =>
-      if d.name.value.nonEmpty then
-        val annots = extractAnnotations(d.mods)
-        buf += SymbolInfo(d.name.value, SymbolKind.Given, file, d.pos.startLine + 1, pkg, Nil, s"given ${d.name.value}", annots)
-    case d: Defn.GivenAlias =>
-      if d.name.value.nonEmpty then
-        val sig = s"given ${d.name.value}: ${d.decltpe.toString()}"
-        val annots = extractAnnotations(d.mods)
-        buf += SymbolInfo(d.name.value, SymbolKind.Given, file, d.pos.startLine + 1, pkg, Nil, sig, annots)
-    case d: Defn.Type =>
-      val sig = s"type ${d.name.value} = ${d.body.toString().take(60)}"
-      val annots = extractAnnotations(d.mods)
-      buf += SymbolInfo(d.name.value, SymbolKind.Type, file, d.pos.startLine + 1, pkg, Nil, sig, annots)
-    case d: Defn.Def =>
-      val params = d.paramClauses.map(_.values.map(p => s"${p.name.value}: ${p.decltpe.map(_.toString()).getOrElse("?")}").mkString(", ")).mkString("(", ")(", ")")
-      val ret = d.decltpe.map(t => s": ${t.toString()}").getOrElse("")
-      val sig = s"def ${d.name.value}$params$ret"
-      val annots = extractAnnotations(d.mods)
-      buf += SymbolInfo(d.name.value, SymbolKind.Def, file, d.pos.startLine + 1, pkg, Nil, sig, annots)
-    case d: Defn.Val =>
-      val annots = extractAnnotations(d.mods)
-      d.pats.foreach {
-        case Pat.Var(name) =>
-          val tpe = d.decltpe.map(t => s": ${t.toString()}").getOrElse("")
-          buf += SymbolInfo(name.value, SymbolKind.Val, file, d.pos.startLine + 1, pkg, Nil, s"val ${name.value}$tpe", annots)
-        case _ =>
-      }
-    case d: Defn.ExtensionGroup =>
-      val recv = d.paramClauses.headOption.flatMap(_.values.headOption).map(p =>
-        s"(${p.name.value}: ${p.decltpe.map(_.toString()).getOrElse("?")})"
-      ).getOrElse("")
-      buf += SymbolInfo("<extension>", SymbolKind.Extension, file, d.pos.startLine + 1, pkg, Nil, s"extension $recv")
-    case _ =>
-
-  def traverse(t: Tree): Unit =
-    visit(t)
-    t.children.foreach(traverse)
-
-  traverse(tree)
-  (buf.toList, bloom, imports, aliases, false)
-
-// ── Source parsing helper ────────────────────────────────────────────────────
-
-def parseFile(path: Path): Option[Source] =
-  val source = try Files.readString(path) catch
-    case _: Exception => return None
-  val input = Input.VirtualFile(path.toString, source)
-  try
-    given scala.meta.Dialect = scala.meta.dialects.Scala3
-    Some(input.parse[Source].get)
-  catch
-    case _: Exception =>
-      try
-        given scala.meta.Dialect = scala.meta.dialects.Scala213
-        Some(input.parse[Source].get)
-      catch
-        case _: Exception => None
-
-// ── Member extraction ───────────────────────────────────────────────────────
-
-case class MemberInfo(name: String, kind: SymbolKind, line: Int, signature: String = "", annotations: List[String] = Nil)
-
-case class BodyInfo(ownerName: String, symbolName: String, sourceText: String, startLine: Int, endLine: Int)
-
-case class HierarchyNode(name: String, kind: Option[SymbolKind], file: Option[Path], line: Option[Int], packageName: String, isExternal: Boolean)
-case class HierarchyTree(root: HierarchyNode, parents: List[HierarchyTree], children: List[HierarchyTree])
-
-case class OverrideInfo(file: Path, line: Int, enclosingClass: String, enclosingKind: SymbolKind, signature: String, packageName: String)
-
-case class ScopeInfo(name: String, kind: String, line: Int)
-
-case class DiffSymbol(name: String, kind: SymbolKind, file: String, line: Int, packageName: String, signature: String)
-
-case class TestCaseInfo(name: String, line: Int, suiteName: String, suiteFile: Path)
-case class TestSuiteInfo(name: String, file: Path, line: Int, tests: List[TestCaseInfo])
-
-def extractMembers(file: Path, symbolName: String): List[MemberInfo] =
-  parseFile(file) match
-    case None => Nil
-    case Some(tree) =>
-      val buf = mutable.ListBuffer.empty[MemberInfo]
-
-      def extractFromTemplate(templ: Template): Unit =
-        templ.stats.foreach {
-          case d: Defn.Def =>
-            val params = d.paramClauses.map(_.values.map(p => s"${p.name.value}: ${p.decltpe.map(_.toString()).getOrElse("?")}").mkString(", ")).mkString("(", ")(", ")")
-            val ret = d.decltpe.map(t => s": ${t.toString()}").getOrElse("")
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Def, d.pos.startLine + 1, s"def ${d.name.value}$params$ret", annots)
-          case d: Defn.Val =>
-            val annots = extractAnnotations(d.mods)
-            d.pats.foreach {
-              case Pat.Var(name) =>
-                val tpe = d.decltpe.map(t => s": ${t.toString()}").getOrElse("")
-                buf += MemberInfo(name.value, SymbolKind.Val, d.pos.startLine + 1, s"val ${name.value}$tpe", annots)
-              case _ =>
-            }
-          case d: Defn.Var =>
-            val annots = extractAnnotations(d.mods)
-            d.pats.foreach {
-              case Pat.Var(name) =>
-                val tpe = d.decltpe.map(t => s": ${t.toString()}").getOrElse("")
-                buf += MemberInfo(name.value, SymbolKind.Var, d.pos.startLine + 1, s"var ${name.value}$tpe", annots)
-              case _ =>
-            }
-          case d: Defn.Type =>
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Type, d.pos.startLine + 1, s"type ${d.name.value} = ${d.body.toString().take(60)}", annots)
-          case d: Decl.Def =>
-            val params = d.paramClauses.map(_.values.map(p => s"${p.name.value}: ${p.decltpe.map(_.toString()).getOrElse("?")}").mkString(", ")).mkString("(", ")(", ")")
-            val ret = s": ${d.decltpe.toString()}"
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Def, d.pos.startLine + 1, s"def ${d.name.value}$params$ret", annots)
-          case d: Decl.Val =>
-            val annots = extractAnnotations(d.mods)
-            d.pats.foreach {
-              case p: Pat.Var =>
-                val tpe = s": ${d.decltpe.toString()}"
-                buf += MemberInfo(p.name.value, SymbolKind.Val, d.pos.startLine + 1, s"val ${p.name.value}$tpe", annots)
-              case _ =>
-            }
-          case d: Decl.Type =>
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Type, d.pos.startLine + 1, s"type ${d.name.value}", annots)
-          case d: Defn.Class =>
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Class, d.pos.startLine + 1, s"class ${d.name.value}", annots)
-          case d: Defn.Trait =>
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Trait, d.pos.startLine + 1, s"trait ${d.name.value}", annots)
-          case d: Defn.Object =>
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Object, d.pos.startLine + 1, s"object ${d.name.value}", annots)
-          case d: Defn.Enum =>
-            val annots = extractAnnotations(d.mods)
-            buf += MemberInfo(d.name.value, SymbolKind.Enum, d.pos.startLine + 1, s"enum ${d.name.value}", annots)
-          case _ =>
-        }
-
-      def findAndExtract(t: Tree): Unit = t match
-        case d: Defn.Class if d.name.value == symbolName => extractFromTemplate(d.templ)
-        case d: Defn.Trait if d.name.value == symbolName => extractFromTemplate(d.templ)
-        case d: Defn.Object if d.name.value == symbolName => extractFromTemplate(d.templ)
-        case d: Defn.Enum if d.name.value == symbolName => extractFromTemplate(d.templ)
-        case _ => t.children.foreach(findAndExtract)
-
-      findAndExtract(tree)
-      buf.toList
-
-// ── Scaladoc extraction ─────────────────────────────────────────────────────
-
-def extractScaladoc(file: Path, targetLine: Int): Option[String] =
-  val lines = try Files.readAllLines(file).asScala.toArray catch
-    case _: Exception => return None
-  // targetLine is 1-indexed, array is 0-indexed
-  var i = targetLine - 2 // line before the symbol
-  // skip blank lines between doc and symbol
-  while i >= 0 && lines(i).trim.isEmpty do i -= 1
-  if i < 0 then return None
-  // Check if this line ends a scaladoc
-  val endLine = i
-  if lines(endLine).trim == "*/" || lines(endLine).trim.endsWith("*/") then
-    // Multi-line or single-line: find the opening /**
-    while i >= 0 && !lines(i).trim.startsWith("/**") do i -= 1
-    if i >= 0 then Some((i to endLine).map(lines(_)).mkString("\n"))
-    else None
-  else if lines(endLine).trim.startsWith("/**") && lines(endLine).trim.endsWith("*/") then
-    // Single-line /** brief */
-    Some(lines(endLine))
-  else None
-
-// ── Test extraction ─────────────────────────────────────────────────────
-
-private val testFnNames = Set("test", "it", "describe")
-
-private def extractTestName(t: Tree): Option[(name: String, line: Int)] =
-  t match
-    case app: Term.Apply =>
-      app.fun match
-        case innerApp: Term.Apply =>
-          innerApp.fun match
-            case fn: Term.Name if testFnNames.contains(fn.value) =>
-              innerApp.argClause.values.collectFirst { case lit: Lit.String => (lit.value, app.pos.startLine + 1) }
-            case _ => None
-        case _ => None
-    case infix: Term.ApplyInfix =>
-      if infix.op.value == "in" || infix.op.value == ">>" then
-        infix.lhs match
-          case lit: Lit.String => Some((lit.value, infix.pos.startLine + 1))
-          case _ => None
-      else None
-    case _ => None
-
-def extractTests(file: Path): List[TestSuiteInfo] = {
-  parseFile(file) match
-    case None => Nil
-    case Some(tree) =>
-      val suites = mutable.ListBuffer.empty[TestSuiteInfo]
-
-      def collectTests(stats: List[Tree], suiteName: String): List[TestCaseInfo] = {
-        val tests = mutable.ListBuffer.empty[TestCaseInfo]
-        def visit(t: Tree): Unit = {
-          extractTestName(t) match
-            case Some((name, line)) =>
-              tests += TestCaseInfo(name, line, suiteName, file)
-            case None =>
-          t.children.foreach(visit)
-        }
-        stats.foreach(visit)
-        tests.toList
-      }
-
-      def findSuites(t: Tree): Unit = {
-        t match
-          case d: Defn.Class =>
-            val tests = collectTests(d.templ.stats, d.name.value)
-            if tests.nonEmpty then
-              suites += TestSuiteInfo(d.name.value, file, d.pos.startLine + 1, tests)
-          case d: Defn.Object =>
-            val tests = collectTests(d.templ.stats, d.name.value)
-            if tests.nonEmpty then
-              suites += TestSuiteInfo(d.name.value, file, d.pos.startLine + 1, tests)
-          case _ =>
-        t.children.foreach(findSuites)
-      }
-
-      findSuites(tree)
-      suites.toList
-}
-
-// ── Body extraction ─────────────────────────────────────────────────────────
-
-def extractBody(file: Path, symbolName: String, ownerName: Option[String]): List[BodyInfo] = {
-  val lines = try Files.readAllLines(file).asScala.toArray catch
-    case _: Exception => return Nil
-  parseFile(file) match
-    case None => Nil
-    case Some(tree) =>
-      val buf = mutable.ListBuffer.empty[BodyInfo]
-
-      def extractFromTree(t: Tree, currentOwner: String): Unit = {
-        t match
-          case d: Defn.Def if d.name.value == symbolName =>
-            if ownerName.isEmpty || ownerName.contains(currentOwner) then
-              val sl = d.pos.startLine
-              val el = d.pos.endLine
-              val body = (sl to el).map(lines(_)).mkString("\n")
-              buf += BodyInfo(currentOwner, d.name.value, body, sl + 1, el + 1)
-          case d: Defn.Val =>
-            d.pats.foreach {
-              case Pat.Var(name) if name.value == symbolName =>
-                if ownerName.isEmpty || ownerName.contains(currentOwner) then
-                  val sl = d.pos.startLine
-                  val el = d.pos.endLine
-                  val body = (sl to el).map(lines(_)).mkString("\n")
-                  buf += BodyInfo(currentOwner, name.value, body, sl + 1, el + 1)
-              case _ =>
-            }
-          case d: Defn.Var =>
-            d.pats.foreach {
-              case Pat.Var(name) if name.value == symbolName =>
-                if ownerName.isEmpty || ownerName.contains(currentOwner) then
-                  val sl = d.pos.startLine
-                  val el = d.pos.endLine
-                  val body = (sl to el).map(lines(_)).mkString("\n")
-                  buf += BodyInfo(currentOwner, name.value, body, sl + 1, el + 1)
-              case _ =>
-            }
-          case d: Defn.Type if d.name.value == symbolName =>
-            if ownerName.isEmpty || ownerName.contains(currentOwner) then
-              val sl = d.pos.startLine
-              val el = d.pos.endLine
-              val body = (sl to el).map(lines(_)).mkString("\n")
-              buf += BodyInfo(currentOwner, d.name.value, body, sl + 1, el + 1)
-          case d: Defn.Class =>
-            if d.name.value == symbolName && (ownerName.isEmpty || ownerName.contains(currentOwner)) then
-              val sl = d.pos.startLine
-              val el = d.pos.endLine
-              val body = (sl to el).map(lines(_)).mkString("\n")
-              buf += BodyInfo(currentOwner, d.name.value, body, sl + 1, el + 1)
-            d.templ.stats.foreach(s => extractFromTree(s, d.name.value))
-          case d: Defn.Trait =>
-            if d.name.value == symbolName && (ownerName.isEmpty || ownerName.contains(currentOwner)) then
-              val sl = d.pos.startLine
-              val el = d.pos.endLine
-              val body = (sl to el).map(lines(_)).mkString("\n")
-              buf += BodyInfo(currentOwner, d.name.value, body, sl + 1, el + 1)
-            d.templ.stats.foreach(s => extractFromTree(s, d.name.value))
-          case d: Defn.Object =>
-            if d.name.value == symbolName && (ownerName.isEmpty || ownerName.contains(currentOwner)) then
-              val sl = d.pos.startLine
-              val el = d.pos.endLine
-              val body = (sl to el).map(lines(_)).mkString("\n")
-              buf += BodyInfo(currentOwner, d.name.value, body, sl + 1, el + 1)
-            d.templ.stats.foreach(s => extractFromTree(s, d.name.value))
-          case d: Defn.Enum =>
-            if d.name.value == symbolName && (ownerName.isEmpty || ownerName.contains(currentOwner)) then
-              val sl = d.pos.startLine
-              val el = d.pos.endLine
-              val body = (sl to el).map(lines(_)).mkString("\n")
-              buf += BodyInfo(currentOwner, d.name.value, body, sl + 1, el + 1)
-            d.templ.stats.foreach(s => extractFromTree(s, d.name.value))
-          case app: Term.Apply =>
-            extractTestName(app) match
-              case Some((name, _)) if name == symbolName =>
-                if ownerName.isEmpty || ownerName.contains(currentOwner) then
-                  val sl = app.pos.startLine
-                  val el = app.pos.endLine
-                  val body = (sl to el).map(lines(_)).mkString("\n")
-                  buf += BodyInfo(currentOwner, name, body, sl + 1, el + 1)
-              case _ =>
-            app.children.foreach(c => extractFromTree(c, currentOwner))
-          case infix: Term.ApplyInfix =>
-            extractTestName(infix) match
-              case Some((name, _)) if name == symbolName =>
-                if ownerName.isEmpty || ownerName.contains(currentOwner) then
-                  val sl = infix.pos.startLine
-                  val el = infix.pos.endLine
-                  val body = (sl to el).map(lines(_)).mkString("\n")
-                  buf += BodyInfo(currentOwner, name, body, sl + 1, el + 1)
-              case _ =>
-            infix.children.foreach(c => extractFromTree(c, currentOwner))
-          case p: Pkg =>
-            p.stats.foreach(s => extractFromTree(s, currentOwner))
-          case _ =>
-      }
-
-      tree.children.foreach(c => extractFromTree(c, ""))
-      buf.toList
-}
-
-// ── Hierarchy building ──────────────────────────────────────────────────────
-
-def buildHierarchy(idx: WorkspaceIndex, symbolName: String, goUp: Boolean, goDown: Boolean, workspace: Path): Option[HierarchyTree] = {
-  val defs = idx.findDefinition(symbolName)
-  if defs.isEmpty then return None
-
-  val sym = defs.head
-  val rootNode = HierarchyNode(sym.name, Some(sym.kind), Some(sym.file), Some(sym.line), sym.packageName, isExternal = false)
-
-  def walkUp(name: String, visited: Set[String]): List[HierarchyTree] = {
-    if visited.contains(name.toLowerCase) then return Nil
-    val newVisited = visited + name.toLowerCase
-    val defs = idx.findDefinition(name)
-    if defs.isEmpty then Nil
-    else {
-      val s = defs.head
-      s.parents.map { parentName =>
-        val parentDefs = idx.findDefinition(parentName)
-        if parentDefs.isEmpty then {
-          val extNode = HierarchyNode(parentName, None, None, None, "", isExternal = true)
-          HierarchyTree(extNode, Nil, Nil)
-        } else {
-          val pd = parentDefs.head
-          val pNode = HierarchyNode(pd.name, Some(pd.kind), Some(pd.file), Some(pd.line), pd.packageName, isExternal = false)
-          val grandParents = walkUp(pd.name, newVisited)
-          HierarchyTree(pNode, grandParents, Nil)
-        }
-      }
-    }
-  }
-
-  def walkDown(name: String, visited: Set[String]): List[HierarchyTree] = {
-    if visited.contains(name.toLowerCase) then return Nil
-    val newVisited = visited + name.toLowerCase
-    val impls = idx.findImplementations(name)
-    impls.map { s =>
-      val node = HierarchyNode(s.name, Some(s.kind), Some(s.file), Some(s.line), s.packageName, isExternal = false)
-      val grandChildren = walkDown(s.name, newVisited)
-      HierarchyTree(node, Nil, grandChildren)
-    }
-  }
-
-  val parents = if goUp then walkUp(sym.name, Set(sym.name.toLowerCase)) else Nil
-  val children = if goDown then walkDown(sym.name, Set(sym.name.toLowerCase)) else Nil
-  Some(HierarchyTree(rootNode, parents, children))
-}
-
-// ── Override finding ────────────────────────────────────────────────────────
-
-def findOverrides(idx: WorkspaceIndex, methodName: String, ofTrait: Option[String], limit: Int): List[OverrideInfo] = {
-  val buf = mutable.ListBuffer.empty[OverrideInfo]
-  val typeKinds = Set(SymbolKind.Class, SymbolKind.Trait, SymbolKind.Object, SymbolKind.Enum)
-
-  ofTrait match
-    case Some(traitName) =>
-      val impls = idx.findImplementations(traitName)
-      impls.foreach { s =>
-        if typeKinds.contains(s.kind) then {
-          val members = extractMembers(s.file, s.name)
-          members.filter(_.name == methodName).foreach { m =>
-            buf += OverrideInfo(s.file, m.line, s.name, s.kind, m.signature, s.packageName)
-          }
-        }
-      }
-    case None =>
-      // Find the method's defining type first
-      val methodDefs = idx.findDefinition(methodName).filter(s => s.kind == SymbolKind.Def || s.kind == SymbolKind.Val)
-      // Also search all types that have a member with this name
-      val allTypes = idx.symbols.filter(s => typeKinds.contains(s.kind))
-      val iter = allTypes.iterator
-      while iter.hasNext && buf.size < limit do {
-        val s = iter.next()
-        val members = extractMembers(s.file, s.name)
-        members.filter(_.name == methodName).foreach { m =>
-          buf += OverrideInfo(s.file, m.line, s.name, s.kind, m.signature, s.packageName)
-        }
-      }
-
-  buf.toList
-}
-
-// ── Scope extraction (context) ──────────────────────────────────────────────
-
-def extractScopes(file: Path, targetLine: Int): List[ScopeInfo] = {
-  parseFile(file) match
-    case None => Nil
-    case Some(tree) =>
-      val buf = mutable.ListBuffer.empty[ScopeInfo]
-
-      def visit(t: Tree): Unit = {
-        val startLine = t.pos.startLine + 1
-        val endLine = t.pos.endLine + 1
-        if targetLine >= startLine && targetLine <= endLine then {
-          t match
-            case d: Pkg =>
-              buf += ScopeInfo(d.ref.toString(), "package", startLine)
-            case d: Defn.Class =>
-              buf += ScopeInfo(d.name.value, "class", startLine)
-            case d: Defn.Trait =>
-              buf += ScopeInfo(d.name.value, "trait", startLine)
-            case d: Defn.Object =>
-              buf += ScopeInfo(d.name.value, "object", startLine)
-            case d: Defn.Enum =>
-              buf += ScopeInfo(d.name.value, "enum", startLine)
-            case d: Defn.Def =>
-              buf += ScopeInfo(d.name.value, "def", startLine)
-            case d: Defn.Val =>
-              d.pats.foreach {
-                case Pat.Var(name) => buf += ScopeInfo(name.value, "val", startLine)
-                case _ =>
-              }
-            case d: Defn.ExtensionGroup =>
-              buf += ScopeInfo("<extension>", "extension", startLine)
-            case _ =>
-          t.children.foreach(visit)
-        }
-      }
-
-      visit(tree)
-      buf.toList
-}
-
-// ── Dependency extraction ───────────────────────────────────────────────────
-
-case class DepInfo(name: String, kind: String, file: Option[Path], line: Option[Int], packageName: String)
-
-def extractDeps(idx: WorkspaceIndex, symbolName: String, workspace: Path): (importDeps: List[DepInfo], bodyDeps: List[DepInfo]) = {
-  val defs = idx.findDefinition(symbolName)
-  if defs.isEmpty then return (Nil, Nil)
-
-  val sym = defs.head
-  val importDeps = mutable.ListBuffer.empty[DepInfo]
-  val bodyDeps = mutable.ListBuffer.empty[DepInfo]
-  val seenNames = mutable.HashSet.empty[String]
-
-  parseFile(sym.file) match
-    case None => (Nil, Nil)
-    case Some(tree) =>
-      // Find the target symbol's AST node and extract info
-      def findNode(t: Tree): Option[Tree] = {
-        t match
-          case d: Defn.Class if d.name.value == symbolName => Some(d)
-          case d: Defn.Trait if d.name.value == symbolName => Some(d)
-          case d: Defn.Object if d.name.value == symbolName => Some(d)
-          case d: Defn.Enum if d.name.value == symbolName => Some(d)
-          case d: Defn.Def if d.name.value == symbolName => Some(d)
-          case _ =>
-            var result: Option[Tree] = None
-            t.children.foreach { c =>
-              if result.isEmpty then result = findNode(c)
-            }
-            result
-      }
-
-      // Collect imports at the file level
-      def collectImports(t: Tree): Unit = {
-        t match
-          case i: Import =>
-            i.importers.foreach { importer =>
-              importer.importees.foreach {
-                case importee: Importee.Name =>
-                  val iname = importee.name.value
-                  if !seenNames.contains(iname) then {
-                    seenNames += iname
-                    val found = idx.findDefinition(iname)
-                    if found.nonEmpty then {
-                      val f = found.head
-                      importDeps += DepInfo(f.name, f.kind.toString.toLowerCase, Some(f.file), Some(f.line), f.packageName)
-                    }
-                  }
-                case importee: Importee.Rename =>
-                  val iname = importee.name.value
-                  if !seenNames.contains(iname) then {
-                    seenNames += iname
-                    val found = idx.findDefinition(iname)
-                    if found.nonEmpty then {
-                      val f = found.head
-                      importDeps += DepInfo(f.name, f.kind.toString.toLowerCase, Some(f.file), Some(f.line), f.packageName)
-                    }
-                  }
-                case _ =>
-              }
-            }
-          case _ =>
-        t.children.foreach(collectImports)
-      }
-      collectImports(tree)
-
-      // Collect type/term references in the symbol's body
-      findNode(tree).foreach { node =>
-        def collectRefs(t: Tree): Unit = {
-          t match
-            case Type.Name(name) if name != symbolName && !seenNames.contains(name) =>
-              seenNames += name
-              val found = idx.findDefinition(name)
-              if found.nonEmpty then {
-                val f = found.head
-                bodyDeps += DepInfo(f.name, f.kind.toString.toLowerCase, Some(f.file), Some(f.line), f.packageName)
-              }
-            case Term.Name(name) if name != symbolName && !seenNames.contains(name) =>
-              seenNames += name
-              val found = idx.findDefinition(name)
-              if found.nonEmpty then {
-                val f = found.head
-                bodyDeps += DepInfo(f.name, f.kind.toString.toLowerCase, Some(f.file), Some(f.line), f.packageName)
-              }
-            case _ =>
-          t.children.foreach(collectRefs)
-        }
-        collectRefs(node)
-      }
-
-      (importDeps.toList, bodyDeps.toList)
-}
-
-// ── Diff extraction ─────────────────────────────────────────────────────────
-
-def runGitDiff(workspace: Path, ref: String): List[String] = {
-  val pb = ProcessBuilder("git", "diff", "--name-only", ref)
-  pb.directory(workspace.toFile)
-  pb.redirectErrorStream(true)
-  val proc = pb.start()
-  val reader = BufferedReader(InputStreamReader(proc.getInputStream))
-  val files = reader.lines().iterator().asScala.filter(_.endsWith(".scala")).toList
-  proc.waitFor()
-  files
-}
-
-def gitShowFile(workspace: Path, ref: String, relPath: String): Option[String] = {
-  try {
-    val pb = ProcessBuilder("git", "show", s"$ref:$relPath")
-    pb.directory(workspace.toFile)
-    pb.redirectErrorStream(false)
-    val proc = pb.start()
-    val reader = BufferedReader(InputStreamReader(proc.getInputStream))
-    val content = reader.lines().iterator().asScala.mkString("\n")
-    val exitCode = proc.waitFor()
-    if exitCode == 0 then Some(content) else None
-  } catch {
-    case _: Exception => None
-  }
-}
-
-def extractSymbolsFromSource(source: String, filePath: String): List[DiffSymbol] = {
-  val input = Input.VirtualFile(filePath, source)
-  val tree = try {
-    given scala.meta.Dialect = scala.meta.dialects.Scala3
-    input.parse[Source].get
-  } catch {
-    case _: Exception =>
-      try {
-        given scala.meta.Dialect = scala.meta.dialects.Scala213
-        input.parse[Source].get
-      } catch {
-        case _: Exception => return Nil
-      }
-  }
-
-  val pkg = tree.children.collectFirst { case p: Pkg => p.ref.toString() }.getOrElse("")
-  val buf = mutable.ListBuffer.empty[DiffSymbol]
-
-  def visit(t: Tree): Unit = {
-    t match
-      case d: Defn.Class =>
-        buf += DiffSymbol(d.name.value, SymbolKind.Class, filePath, d.pos.startLine + 1, pkg, s"class ${d.name.value}")
-      case d: Defn.Trait =>
-        buf += DiffSymbol(d.name.value, SymbolKind.Trait, filePath, d.pos.startLine + 1, pkg, s"trait ${d.name.value}")
-      case d: Defn.Object =>
-        buf += DiffSymbol(d.name.value, SymbolKind.Object, filePath, d.pos.startLine + 1, pkg, s"object ${d.name.value}")
-      case d: Defn.Enum =>
-        buf += DiffSymbol(d.name.value, SymbolKind.Enum, filePath, d.pos.startLine + 1, pkg, s"enum ${d.name.value}")
-      case d: Defn.Def =>
-        buf += DiffSymbol(d.name.value, SymbolKind.Def, filePath, d.pos.startLine + 1, pkg, s"def ${d.name.value}")
-      case d: Defn.Val =>
-        d.pats.foreach {
-          case Pat.Var(name) =>
-            buf += DiffSymbol(name.value, SymbolKind.Val, filePath, d.pos.startLine + 1, pkg, s"val ${name.value}")
-          case _ =>
-        }
-      case d: Defn.Type =>
-        buf += DiffSymbol(d.name.value, SymbolKind.Type, filePath, d.pos.startLine + 1, pkg, s"type ${d.name.value}")
-      case _ =>
-  }
-
-  def traverse(t: Tree): Unit = {
-    visit(t)
-    t.children.foreach(traverse)
-  }
-
-  traverse(tree)
-  buf.toList
-}
-
-// ── AST pattern matching ────────────────────────────────────────────────────
-
-case class AstPatternMatch(name: String, kind: SymbolKind, file: Path, line: Int, packageName: String, signature: String)
-
-def astPatternSearch(idx: WorkspaceIndex, workspace: Path,
-                     hasMethod: Option[String], extendsTrait: Option[String],
-                     bodyContains: Option[String], noTests: Boolean,
-                     pathFilter: Option[String], limit: Int): List[AstPatternMatch] = {
-  val typeKinds = Set(SymbolKind.Class, SymbolKind.Trait, SymbolKind.Object, SymbolKind.Enum)
-  var candidates = idx.symbols.filter(s => typeKinds.contains(s.kind))
-  if noTests then candidates = candidates.filter(s => !isTestFile(s.file, workspace))
-  pathFilter.foreach { p => candidates = candidates.filter(s => matchesPath(s.file, p, workspace)) }
-
-  // Filter by extends
-  extendsTrait.foreach { traitName =>
-    candidates = candidates.filter(_.parents.exists(_.equalsIgnoreCase(traitName)))
-  }
-
-  val buf = mutable.ListBuffer.empty[AstPatternMatch]
-
-  val iter = candidates.iterator
-  while iter.hasNext && buf.size < limit do {
-    val s = iter.next()
-    var matches = true
-
-    // Filter by has-method
-    hasMethod.foreach { methodName =>
-      val members = extractMembers(s.file, s.name)
-      if !members.exists(_.name == methodName) then matches = false
-    }
-
-    // Filter by body-contains
-    bodyContains.foreach { pattern =>
-      if matches then {
-        val bodies = extractBody(s.file, s.name, None)
-        if !bodies.exists(_.sourceText.contains(pattern)) then matches = false
-      }
-    }
-
-    if matches then {
-      buf += AstPatternMatch(s.name, s.kind, s.file, s.line, s.packageName, s.signature)
-    }
-  }
-  buf.toList
-}
-
-// ── Binary persistence ──────────────────────────────────────────────────────
-
-object IndexPersistence:
-  private val MAGIC = 0x53584458
-  private val VERSION: Byte = 6
-
-  def indexPath(workspace: Path): Path = workspace.resolve(".scalex").resolve("index.bin")
-
-  def save(workspace: Path, files: List[IndexedFile]): Unit =
-    val dir = workspace.resolve(".scalex")
-    if !Files.exists(dir) then Files.createDirectories(dir)
-
-    val stringTable = mutable.LinkedHashMap.empty[String, Int]
-    def intern(s: String): Int =
-      stringTable.getOrElseUpdate(s, stringTable.size)
-
-    files.foreach { f =>
-      intern(f.relativePath)
-      intern(f.oid)
-      f.symbols.foreach { s =>
-        intern(s.name)
-        intern(s.packageName)
-        intern(s.signature)
-        s.parents.foreach(intern)
-        s.annotations.foreach(intern)
-      }
-      f.imports.foreach(intern)
-      f.aliases.foreach { (k, v) => intern(k); intern(v) }
-    }
-
-    val out = DataOutputStream(BufferedOutputStream(Files.newOutputStream(indexPath(workspace)), 1 << 16))
-    try
-      out.writeInt(MAGIC)
-      out.writeByte(VERSION)
-
-      val strings = stringTable.keys.toArray
-      out.writeInt(strings.length)
-      strings.foreach(out.writeUTF)
-
-      out.writeInt(files.size)
-      files.foreach { f =>
-        out.writeInt(intern(f.relativePath))
-        out.writeInt(intern(f.oid))
-
-        out.writeShort(f.symbols.size)
-        f.symbols.foreach { s =>
-          out.writeInt(intern(s.name))
-          out.writeByte(s.kind.id)
-          out.writeInt(s.line)
-          out.writeInt(intern(s.packageName))
-          out.writeInt(intern(s.signature))
-          out.writeShort(s.parents.size)
-          s.parents.foreach(p => out.writeInt(intern(p)))
-          out.writeShort(s.annotations.size)
-          s.annotations.foreach(a => out.writeInt(intern(a)))
-        }
-
-        // Imports
-        out.writeShort(f.imports.size)
-        f.imports.foreach(i => out.writeInt(intern(i)))
-
-        // Aliases
-        out.writeShort(f.aliases.size)
-        f.aliases.foreach { (k, v) =>
-          out.writeInt(intern(k))
-          out.writeInt(intern(v))
-        }
-
-        // Bloom filter
-        val bloomBytes = java.io.ByteArrayOutputStream()
-        f.identifierBloom.writeTo(bloomBytes)
-        val ba = bloomBytes.toByteArray
-        out.writeInt(ba.length)
-        out.write(ba)
-
-        // Parse failed flag
-        out.writeBoolean(f.parseFailed)
-      }
-    finally out.close()
-
-  def load(workspace: Path, loadBlooms: Boolean = true): Option[Map[String, IndexedFile]] =
-    val p = indexPath(workspace)
-    if !Files.exists(p) then return None
-
-    try
-      val in = DataInputStream(BufferedInputStream(Files.newInputStream(p), 1 << 16))
-      try
-        val magic = in.readInt()
-        if magic != MAGIC then return None
-        val version = in.readByte()
-        if version != VERSION then return None
-
-        val strCount = in.readInt()
-        val strings = Array.fill(strCount)(in.readUTF())
-
-        val fileCount = in.readInt()
-        val result = mutable.HashMap.empty[String, IndexedFile]
-
-        var fi = 0
-        while fi < fileCount do
-          val relPath = strings(in.readInt())
-          val oid = strings(in.readInt())
-
-          val symCount = in.readShort()
-          val syms = List.newBuilder[SymbolInfo]
-          var si = 0
-          while si < symCount do
-            val name = strings(in.readInt())
-            val kind = SymbolKind.fromId(in.readByte())
-            val line = in.readInt()
-            val pkg = strings(in.readInt())
-            val sig = strings(in.readInt())
-            val parentCount = in.readShort()
-            val parents = (0 until parentCount).map(_ => strings(in.readInt())).toList
-            val annotCount = in.readShort()
-            val annots = (0 until annotCount).map(_ => strings(in.readInt())).toList
-            syms += SymbolInfo(name, kind, workspace.resolve(relPath), line, pkg, parents, sig, annots)
-            si += 1
-
-          // Imports
-          val importCount = in.readShort()
-          val imports = (0 until importCount).map(_ => strings(in.readInt())).toList
-
-          // Aliases
-          val aliasCount = in.readShort()
-          val aliases = (0 until aliasCount).map { _ =>
-            val k = strings(in.readInt())
-            val v = strings(in.readInt())
-            k -> v
-          }.toMap
-
-          // Bloom filter
-          val bloomLen = in.readInt()
-          val bloom = if loadBlooms then
-            val bloomBytes = new Array[Byte](bloomLen)
-            in.readFully(bloomBytes)
-            BloomFilter.readFrom(
-              java.io.ByteArrayInputStream(bloomBytes),
-              Funnels.unencodedCharsFunnel()
-            )
-          else
-            in.skipBytes(bloomLen)
-            null
-
-          // Parse failed flag
-          val parseFailed = in.readBoolean()
-
-          result(relPath) = IndexedFile(relPath, oid, syms.result(), bloom, imports, aliases, parseFailed)
-          fi += 1
-
-        Some(result.toMap)
-      finally in.close()
-    catch
-      case _: Exception => None
-
-// ── Workspace index ─────────────────────────────────────────────────────────
-
-class WorkspaceIndex(val workspace: Path, val needBlooms: Boolean = true):
-  var symbols: List[SymbolInfo] = Nil
-  var filesByPath: Map[Path, List[SymbolInfo]] = Map.empty
-  var symbolsByName: Map[String, List[SymbolInfo]] = Map.empty
-  var packages: Set[String] = Set.empty
-  var gitFiles: List[GitFile] = Nil
-  private var indexedFiles: List[IndexedFile] = Nil
-  var parentIndex: Map[String, List[SymbolInfo]] = Map.empty
-
-  private var distinctSymbols: List[SymbolInfo] = Nil
-  var packageToSymbols: Map[String, Set[String]] = Map.empty
-  private var indexedByPath: Map[String, IndexedFile] = Map.empty
-  private var aliasIndex: Map[String, List[(file: IndexedFile, alias: String)]] = Map.empty
-  private var annotationIndex: Map[String, List[SymbolInfo]] = Map.empty
-
-  var fileCount: Int = 0
-  var indexTimeMs: Long = 0
-  var parsedCount: Int = 0
-  var skippedCount: Int = 0
-  var parseFailures: Int = 0
-  var parseFailedFiles: List[String] = Nil
-  var cachedLoad: Boolean = false
-
-  def index(): Unit =
-    val t0 = System.nanoTime()
-    gitFiles = gitLsFiles(workspace)
-    fileCount = gitFiles.size
-
-    val cached = IndexPersistence.load(workspace, needBlooms)
-    val result = mutable.ListBuffer.empty[IndexedFile]
-
-    cached match
-      case Some(cachedMap) =>
-        cachedLoad = true
-        val toParseQueue = ConcurrentLinkedQueue[IndexedFile]()
-        val toParse = mutable.ListBuffer.empty[GitFile]
-
-        gitFiles.foreach { gf =>
-          val rel = workspace.relativize(gf.path).toString
-          cachedMap.get(rel) match
-            case Some(cf) if cf.oid == gf.oid =>
-              result += cf
-              skippedCount += 1
-            case _ =>
-              toParse += gf
-        }
-
-        toParse.asJava.parallelStream().forEach { gf =>
-          val rel = workspace.relativize(gf.path).toString
-          val (syms, bloom, imports, aliases, failed) = extractSymbols(gf.path)
-          toParseQueue.add(IndexedFile(rel, gf.oid, syms, bloom, imports, aliases, failed))
-        }
-        result ++= toParseQueue.asScala
-        parsedCount = toParse.size
-
-      case None =>
-        val queue = ConcurrentLinkedQueue[IndexedFile]()
-        gitFiles.asJava.parallelStream().forEach { gf =>
-          val rel = workspace.relativize(gf.path).toString
-          val (syms, bloom, imports, aliases, failed) = extractSymbols(gf.path)
-          queue.add(IndexedFile(rel, gf.oid, syms, bloom, imports, aliases, failed))
-        }
-        result ++= queue.asScala
-        parsedCount = gitFiles.size
-
-    indexedFiles = result.toList
-    parseFailedFiles = indexedFiles.collect {
-      case f if f.parseFailed => f.relativePath
-    }
-    parseFailures = parseFailedFiles.size
-    // Single-pass over symbols: build all symbol-level indexes
-    val allSyms = mutable.ListBuffer.empty[SymbolInfo]
-    val byPath = mutable.HashMap.empty[Path, mutable.ListBuffer[SymbolInfo]]
-    val byName = mutable.HashMap.empty[String, mutable.ListBuffer[SymbolInfo]]
-    val pkgs = mutable.HashSet.empty[String]
-    val pIdx = mutable.HashMap.empty[String, mutable.ListBuffer[SymbolInfo]]
-    val pkgToSyms = mutable.HashMap.empty[String, mutable.HashSet[String]]
-    val distinctSeen = mutable.HashSet.empty[(String, Path, Int)]
-    val distinctBuf = mutable.ListBuffer.empty[SymbolInfo]
-    val aByAnnot = mutable.HashMap.empty[String, mutable.ListBuffer[SymbolInfo]]
-    indexedFiles.foreach { f =>
-      f.symbols.foreach { s =>
-        allSyms += s
-        byPath.getOrElseUpdate(s.file, mutable.ListBuffer.empty) += s
-        byName.getOrElseUpdate(s.name.toLowerCase, mutable.ListBuffer.empty) += s
-        if s.packageName.nonEmpty then pkgs += s.packageName
-        s.parents.foreach { p =>
-          pIdx.getOrElseUpdate(p.toLowerCase, mutable.ListBuffer.empty) += s
-        }
-        s.annotations.foreach { a =>
-          aByAnnot.getOrElseUpdate(a.toLowerCase, mutable.ListBuffer.empty) += s
-        }
-        pkgToSyms.getOrElseUpdate(s.packageName, mutable.HashSet.empty) += s.name
-        val key = (s.name, s.file, s.line)
-        if distinctSeen.add(key) then distinctBuf += s
-      }
-    }
-    symbols = allSyms.toList
-    distinctSymbols = distinctBuf.toList
-    filesByPath = byPath.map((k, v) => k -> v.toList).toMap
-    symbolsByName = byName.map((k, v) => k -> v.toList).toMap
-    packages = pkgs.toSet
-    parentIndex = pIdx.map((k, v) => k -> v.toList).toMap
-    annotationIndex = aByAnnot.map((k, v) => k -> v.toList).toMap
-    packageToSymbols = pkgToSyms.map((k, v) => k -> v.toSet).toMap
-
-    // Single-pass over indexedFiles: build file-level indexes
-    val iByPath = mutable.HashMap.empty[String, IndexedFile]
-    val aIdx = mutable.HashMap.empty[String, mutable.ListBuffer[(file: IndexedFile, alias: String)]]
-    indexedFiles.foreach { f =>
-      iByPath(f.relativePath) = f
-      f.aliases.foreach { (orig, alias) =>
-        aIdx.getOrElseUpdate(orig, mutable.ListBuffer.empty) += ((f, alias))
-      }
-    }
-    indexedByPath = iByPath.toMap
-    aliasIndex = aIdx.map((k, v) => k -> v.toList).toMap
-    indexTimeMs = (System.nanoTime() - t0) / 1_000_000
-
-    if parsedCount > 0 then
-      if !needBlooms then
-        // Reload with blooms for newly parsed files that have null blooms
-        indexedFiles = indexedFiles.map { f =>
-          if f.identifierBloom == null then
-            val source = try Files.readString(workspace.resolve(f.relativePath)) catch
-              case _: Exception => ""
-            f.copy(identifierBloom = buildBloomFilterFromSource(source))
-          else f
-        }
-      IndexPersistence.save(workspace, indexedFiles)
-
-  def findDefinition(name: String): List[SymbolInfo] =
-    symbolsByName.getOrElse(name.toLowerCase, Nil)
-
-  def findImplementations(name: String): List[SymbolInfo] =
-    parentIndex.getOrElse(name.toLowerCase, Nil)
-
-  def findAnnotated(annotation: String): List[SymbolInfo] =
-    annotationIndex.getOrElse(annotation.toLowerCase, Nil)
-
-  def grepFiles(pattern: String, noTests: Boolean, pathFilter: Option[String],
-                timeoutMs: Long = defaultTimeoutMs): (results: List[Reference], timedOut: Boolean) =
-    val regex = try java.util.regex.Pattern.compile(pattern)
-    catch
-      case e: java.util.regex.PatternSyntaxException =>
-        System.err.println(s"Invalid regex: ${e.getMessage}")
-        return (Nil, false)
-    var candidates = gitFiles
-    if noTests then candidates = candidates.filter(gf => !isTestFile(gf.path, workspace))
-    pathFilter.foreach { p => candidates = candidates.filter(gf => matchesPath(gf.path, p, workspace)) }
-    val deadline = System.nanoTime() + timeoutMs * 1_000_000
-    var grepTimedOut = false
-    val results = ConcurrentLinkedQueue[Reference]()
-    candidates.asJava.parallelStream().forEach { gf =>
-      if System.nanoTime() < deadline then {
-        try {
-          val lines = Files.readAllLines(gf.path).asScala
-          lines.zipWithIndex.foreach { case (line, idx) =>
-            if System.nanoTime() < deadline then {
-              if regex.matcher(line).find() then
-                results.add(Reference(gf.path, idx + 1, line.trim))
-            } else grepTimedOut = true
-          }
-        } catch { case _: Exception => () }
-      } else grepTimedOut = true
-    }
-    (results.asScala.toList.sortBy(r => (workspace.relativize(r.file).toString, r.line)), grepTimedOut)
-
-  def search(query: String): List[SymbolInfo] =
-    val lower = query.toLowerCase
-    val exact = mutable.ListBuffer.empty[SymbolInfo]
-    val prefix = mutable.ListBuffer.empty[SymbolInfo]
-    val contains = mutable.ListBuffer.empty[SymbolInfo]
-    val fuzzy = mutable.ListBuffer.empty[SymbolInfo]
-
-    distinctSymbols.foreach { s =>
-      val n = s.name.toLowerCase
-      if n == lower then exact += s
-      else if n.startsWith(lower) then prefix += s
-      else if n.contains(lower) then contains += s
-      else if camelCaseMatch(lower, s.name) then fuzzy += s
-    }
-    exact.toList ++ prefix.toList ++ contains.toList ++ fuzzy.sortBy(_.name.length).toList
-
-  def fileSymbols(path: String): List[SymbolInfo] =
-    val resolved = if Path.of(path).isAbsolute then Path.of(path)
-                   else workspace.resolve(path)
-    filesByPath.getOrElse(resolved, Nil)
-
-  def searchFiles(query: String): List[String] =
-    val lower = query.toLowerCase
-    val exact = mutable.ListBuffer.empty[String]
-    val prefix = mutable.ListBuffer.empty[String]
-    val contains = mutable.ListBuffer.empty[String]
-    val fuzzy = mutable.ListBuffer.empty[String]
-
-    indexedFiles.foreach { f =>
-      val fileName = f.relativePath.substring(f.relativePath.lastIndexOf('/') + 1).stripSuffix(".scala")
-      val n = fileName.toLowerCase
-      if n == lower then exact += f.relativePath
-      else if n.startsWith(lower) then prefix += f.relativePath
-      else if n.contains(lower) then contains += f.relativePath
-      else if camelCaseMatch(lower, fileName) then fuzzy += f.relativePath
-    }
-    exact.toList ++ prefix.toList ++ contains.toList ++ fuzzy.sortBy(_.length).toList
-
-  private val defaultTimeoutMs = 20_000L
-  var timedOut: Boolean = false
-
-  def findReferences(name: String, timeoutMs: Long = defaultTimeoutMs): List[Reference] =
-    val candidates = indexedFiles.filter(f => f.identifierBloom == null || f.identifierBloom.mightContain(name))
-    val aliasFiles = aliasIndex.getOrElse(name, Nil)
-    val candidateSet = candidates.map(_.relativePath).toSet
-    val extraFiles = aliasFiles.collect {
-      case (f, _) if !candidateSet.contains(f.relativePath) => f
-    }
-    val allCandidates = candidates ++ extraFiles
-    val fileAliasMap = aliasFiles.map((f, alias) => f.relativePath -> alias).toMap
-
-    val deadline = System.nanoTime() + timeoutMs * 1_000_000
-    timedOut = false
-    val results = ConcurrentLinkedQueue[Reference]()
-    val seen = java.util.concurrent.ConcurrentHashMap.newKeySet[String]()
-    allCandidates.asJava.parallelStream().forEach { idxFile =>
-      if System.nanoTime() < deadline then
-        val path = workspace.resolve(idxFile.relativePath)
-        val lines = try Files.readAllLines(path).asScala catch
-          case _: Exception => Seq.empty
-        val aliasName = fileAliasMap.get(idxFile.relativePath)
-        lines.zipWithIndex.foreach {
-          case (line, idx) if System.nanoTime() < deadline =>
-            val key = s"${idxFile.relativePath}:${idx + 1}"
-            if containsWord(line, name) && seen.add(key) then
-              results.add(Reference(path, idx + 1, line.trim))
-            else aliasName match
-              case Some(alias) if containsWord(line, alias) && seen.add(key) =>
-                results.add(Reference(path, idx + 1, line.trim, Some(s"via alias $alias")))
-              case _ =>
-          case _ =>
-        }
-      else timedOut = true
-    }
-    results.asScala.toList
-
-  def categorizeReferences(name: String): Map[RefCategory, List[Reference]] =
-    val refs = findReferences(name)
-    refs.groupBy { r =>
-      val line = r.contextLine
-      if line.matches("""^\s*(trait|class|object|enum|given|type|def|val|var)\s+.*""") && containsWord(line, name) &&
-         (line.contains(s"trait $name") || line.contains(s"class $name") || line.contains(s"object $name") ||
-          line.contains(s"enum $name") || line.contains(s"type $name") ||
-          line.matches(s""".*given\\s+\\w*$name.*""")) then
-        RefCategory.Definition
-      else if line.matches(""".*\b(extends|with)\b.*""") && containsWord(line, name) then
-        RefCategory.ExtendedBy
-      else if line.trim.startsWith("import ") then
-        RefCategory.ImportedBy
-      else if line.matches("""^\s*(//|/\*|\*).*""") then
-        RefCategory.Comment
-      else if line.matches(s""".*:\\s*$name.*""") || line.matches(s""".*\\[$name.*""") then
-        RefCategory.UsedAsType
-      else
-        RefCategory.Usage
-    }
-
-  def findImports(name: String, timeoutMs: Long = defaultTimeoutMs): List[Reference] =
-    val candidates = indexedFiles.filter(f => f.identifierBloom == null || f.identifierBloom.mightContain(name))
-    val deadline = System.nanoTime() + timeoutMs * 1_000_000
-    timedOut = false
-    val results = ConcurrentLinkedQueue[Reference]()
-    val resultPaths = java.util.concurrent.ConcurrentHashMap.newKeySet[String]()
-    candidates.asJava.parallelStream().forEach { idxFile =>
-      if System.nanoTime() < deadline then
-        val path = workspace.resolve(idxFile.relativePath)
-        val lines = try Files.readAllLines(path).asScala catch
-          case _: Exception => Seq.empty
-        lines.zipWithIndex.foreach {
-          case (line, idx) if System.nanoTime() < deadline && line.trim.startsWith("import ") && containsWord(line, name) =>
-            results.add(Reference(path, idx + 1, line.trim))
-            resultPaths.add(s"${idxFile.relativePath}:${idx + 1}")
-          case _ =>
-        }
-      else timedOut = true
-    }
-
-    // Also find wildcard imports that resolve to a package containing the target symbol
-    val targetPkgs = symbolsByName.getOrElse(name.toLowerCase, Nil).map(_.packageName).toSet
-    if targetPkgs.nonEmpty then
-      for idxFile <- indexedFiles if System.nanoTime() < deadline do
-        for imp <- idxFile.imports do
-          val trimmed = imp.trim.stripPrefix("import ")
-          if (trimmed.endsWith("._") || trimmed.endsWith(".*")) then
-            val pkg = trimmed.dropRight(2)
-            if targetPkgs.contains(pkg) then
-              val path = workspace.resolve(idxFile.relativePath)
-              try {
-                val lines = Files.readAllLines(path).asScala
-                lines.zipWithIndex.foreach { case (line, lineIdx) =>
-                  if line.trim == imp.trim then
-                    val key = s"${idxFile.relativePath}:${lineIdx + 1}"
-                    if !resultPaths.contains(key) then
-                      results.add(Reference(path, lineIdx + 1, line.trim))
-                      resultPaths.add(key)
-                }
-              } catch { case _: Exception => () }
-
-    results.asScala.toList
-
-  private def filePackage(idxFile: IndexedFile): String =
-    idxFile.symbols.headOption.map(_.packageName).getOrElse("")
-
-  def resolveConfidence(ref: Reference, targetName: String, targetPackages: Set[String]): Confidence =
-    val relPath = workspace.relativize(ref.file).toString
-    indexedByPath.get(relPath) match
-      case None => Confidence.Low
-      case Some(idxFile) =>
-        val filePkg = filePackage(idxFile)
-        if targetPackages.contains(filePkg) then Confidence.High
-        else
-          val imports = idxFile.imports
-          val hasExplicit = imports.exists { imp =>
-            imp.contains(s".$targetName") || imp.contains(s"{$targetName") ||
-            imp.contains(s", $targetName") || imp.contains(s"$targetName,")
-          }
-          val hasAliasMatch = idxFile.aliases.exists { (orig, alias) =>
-            alias == targetName || orig == targetName
-          }
-          if hasExplicit || hasAliasMatch then Confidence.High
-          else
-            val hasWildcard = imports.exists { imp =>
-              val trimmed = imp.trim.stripPrefix("import ")
-              (trimmed.endsWith("._") || trimmed.endsWith(".*")) && {
-                val pkg = trimmed.dropRight(2)
-                targetPackages.contains(pkg)
-              }
-            }
-            if hasWildcard then Confidence.Medium
-            else Confidence.Low
-
-  private def isSegmentStart(name: String, i: Int): Boolean =
-    i == 0 || name(i).isUpper || (i > 0 && name(i - 1) == '_')
-
-  private def camelCaseMatch(query: String, name: String): Boolean =
-    if query.length < 2 then return false
-    val qLower = query.toLowerCase
-    val nLower = name.toLowerCase
-    var qi = 0
-    var ni = 0
-    while qi < qLower.length && ni < nLower.length do
-      if qLower(qi) == nLower(ni) then
-        qi += 1
-        ni += 1
-      else
-        ni += 1
-        while ni < nLower.length && !isSegmentStart(name, ni) do ni += 1
-    qi == qLower.length
-
-  private def containsWord(line: String, word: String): Boolean =
-    var i = line.indexOf(word)
-    while i >= 0 do
-      val before = i == 0 || !line(i - 1).isLetterOrDigit
-      val after = i + word.length >= line.length || !line(i + word.length).isLetterOrDigit
-      if before && after then return true
-      i = line.indexOf(word, i + 1)
-    false
-
-// ── Filtering helpers ────────────────────────────────────────────────────────
-
-def isTestFile(path: Path, workspace: Path): Boolean =
-  val rel = workspace.relativize(path).toString
-  rel.contains("/test/") || rel.contains("/tests/") || rel.contains("/testing/") ||
-  rel.startsWith("bench-") || rel.contains("/bench-") ||
-  rel.endsWith("Test.scala") || rel.endsWith("Spec.scala") || rel.endsWith("Suite.scala") ||
-  rel.endsWith(".test.scala")
-
-def matchesPath(file: Path, prefix: String, workspace: Path): Boolean =
-  val rel = workspace.relativize(file).toString
-  rel.startsWith(prefix)
 
 // ── CLI ─────────────────────────────────────────────────────────────────────
