+++ src/project.scala
@@ -1,0 +1,4 @@
+//> using scala 3.8.2
+//> using dep org.scalameta::scalameta:4.15.2
+//> using dep com.google.guava:guava:33.5.0-jre
+//> using test.dep org.scalameta::munit:1.2.4
